refactor: pro/lite/ts
