--- scaffolds/ice-design-lite/mock/index.js
@@ -1,12 +1,0 @@
-// Refer to https://www.npmjs.com/package/webpack-dev-mock for detail configuration
-module.exports = {
-  'GET /api/profile': {
-    status: 'SUCCESS',
-    data: {
-      name: '淘小宝',
-      department: '技术部',
-      avatar: 'https://img.alicdn.com/tfs/TB1L6tBXQyWBuNjy0FpXXassXXa-80-80.png',
-      userid: 10001,
-    },
-  },
-};
