+++ scaffolds/ice-design-lite/package.json
@@ -29,7 +29,6 @@
     "prop-types": "^15.5.8",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
-    "react-intl": "^2.9.0",
     "react-router-dom": "^5.0.1"
   },
   "devDependencies": {
