--- scaffolds/ice-design-lite/src/dataSourceConfig.js
@@ -1,4 +1,0 @@
-export const userProfile = {
-  url: '/api/profile',
-  method: 'GET',
-};
