+++ scaffolds/ice-design-lite/src/layouts/BasicLayout/components/Aside/index.module.scss
@@ -61,6 +61,6 @@
     }
   }
 }
-.nav-box{
+.nav-box {
   width: 200px;
 }
