+++ scaffolds/ice-design-lite/src/layouts/BasicLayout/index.jsx
@@ -36,7 +36,7 @@ const BasicLayout = (props) => {
 
   useEffect(() => {
     enquireScreenRegister();
-  }, [enquireScreenRegister]);
+  }, []);
 
   const isMobile = isScreen !== 'isDesktop';
