--- scaffolds/ice-design-lite/src/locales/en-US/index.js
@@ -1,5 +1,0 @@
-import menu from './menu';
-
-export default {
-  ...menu,
-};
