--- scaffolds/ice-design-lite/src/locales/en-US/menu.js
@@ -1,21 +1,0 @@
-export default {
-  // Aside Menu
-  'app.menu.dashboard': 'Dashboard',
-  'app.menu.dashboard.monitor': 'Monitor',
-
-  'app.menu.chart': 'Chart',
-  'app.menu.chart.basic': 'Basic Chart',
-  'app.menu.chart.general': 'General Chart',
-
-  'app.menu.table': 'Table',
-  'app.menu.table.basic': 'Basic Table',
-  'app.menu.table.general': 'General Table',
-
-  // Top Header
-  'app.header.feedback': 'Feedback',
-  'app.header.help': 'Help',
-  'app.header.user.name': 'Tao',
-  'app.header.user.department': 'Tech',
-  'app.header.user.setting': 'Setting',
-  'app.header.user.logout': 'Logout',
-};
