--- scaffolds/ice-design-lite/src/locales/zh-CN/index.js
@@ -1,5 +1,0 @@
-import menu from './menu';
-
-export default {
-  ...menu,
-};
