--- scaffolds/ice-design-lite/src/locales/zh-CN/menu.js
@@ -1,19 +1,0 @@
-export default {
-  // Aside Menu
-  'app.menu.dashboard': 'Dashboard',
-  'app.menu.dashboard.monitor': '监控页',
-
-  'app.menu.chart': '图表页',
-  'app.menu.chart.basic': '基础图表',
-  'app.menu.chart.general': '通用图表',
-
-  'app.menu.table': '表格页',
-  'app.menu.table.basic': '基础表格',
-  'app.menu.table.general': '通用表格',
-
-  // Top Header
-  'app.header.feedback': '反馈',
-  'app.header.help': '帮助',
-  'app.header.user.setting': '设置',
-  'app.header.user.logout': '退出',
-};
