--- scaffolds/ice-design-lite/src/stores/index.js
@@ -1,7 +1,0 @@
-import Icestore from '@ice/store';
-import userProfile from './userProfile';
-
-const icestore = new Icestore();
-icestore.registerStore('userProfile', userProfile);
-
-export default icestore;
