+++ scaffolds/ice-design-pro/src/config/dataSource.js
@@ -1,3 +1,4 @@
+
 export const userProfile = {
   url: '/api/profile',
   method: 'GET',
