+++ scaffolds/ice-design-pro/src/index.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-// 引入默认全局样式
-import '@alifd/next/reset.scss';
 import './global.scss';
 
 // 引入基础配置文件
