+++ scaffolds/ice-design-pro/src/layouts/BasicLayout/index.jsx
@@ -34,7 +34,7 @@ export default function BasicLayout(props) {
 
   useEffect(() => {
     enquireScreenRegister();
-  }, [enquireScreenRegister]);
+  }, []);
 
   const isMobile = isScreen !== 'isDesktop';
