+++ scaffolds/ice-typescript-starter/src/global.scss
