+++ scaffolds/ice-typescript-starter/src/index.tsx
@@ -1,7 +1,6 @@
 import ReactDOM from 'react-dom';
 
-// 载入默认全局样式 normalize
-import '@alifd/next/reset.scss';
+import './global.scss';
 
 import router from './router';
